chore: rename logo.tsx
