+++ components/SideNavbar/SideNavbarHeader.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import { useContext } from 'react'
 import { AiOutlineClose } from 'react-icons/ai'
 import { GlobalContext } from 'context/GlobalContext'
-import Logo from '../logo'
+import Logo from '../logo/logo'
 import { IContext } from 'types'
 
 export const SideNavbarHeader: FC = () => {
