+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Head from 'next/head'
-import Logo from 'components/logo'
+import Logo from 'components/logo/logo'
 import TypewriterComponent from 'typewriter-effect'
 import { sidebarData } from '../database/data'
 export default function Home() {
